ACS-6390: Refactoring - moved error handling to common module
