+++ common/src/main/java/org/alfresco/hxi_connector/common/exception/ResourceNotFoundException.java
@@ -2,7 +2,7 @@
  * #%L
  * Alfresco HX Insight Connector
  * %%
- * Copyright (C) 2024 Alfresco Software Limited
+ * Copyright (C) 2023 - 2024 Alfresco Software Limited
  * %%
  * This file is part of the Alfresco software.
  * If the software was purchased under a paid Alfresco license, the terms of
@@ -23,22 +23,12 @@
  * along with Alfresco. If not, see <http://www.gnu.org/licenses/>.
  * #L%
  */
-package org.alfresco.hxi_connector.live_ingester.domain.exception;
+package org.alfresco.hxi_connector.common.exception;
 
-public class EndpointClientErrorException extends RuntimeException
+public class ResourceNotFoundException extends EndpointClientErrorException
 {
-    public EndpointClientErrorException(String message)
+    public ResourceNotFoundException(String message)
     {
         super(message);
-    }
-
-    public EndpointClientErrorException(Throwable cause)
-    {
-        super(cause);
-    }
-
-    public EndpointClientErrorException(String message, Throwable cause)
-    {
-        super(message, cause);
     }
 }
